OneDrive/Documents/Projects/Authentication + User Management service/requirements.txt
fastapi
uvicorn
bcrypt
PyJWT
